+++ common-proxy/src/main/java/de/maxhenkel/voicechat/sniffer/IncompatibleVoiceChatException.java
@@ -1,0 +1,17 @@
+package de.maxhenkel.voicechat.sniffer;
+
+public class IncompatibleVoiceChatException extends Exception{
+
+    public IncompatibleVoiceChatException(String message) {
+        super(message);
+    }
+
+    public IncompatibleVoiceChatException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public IncompatibleVoiceChatException(Throwable cause) {
+        super(cause);
+    }
+
+}