+++ bungeecord/changelog.md
@@ -1,1 +1,2 @@
 - Fixed ping buffer size
+- Added compatibility check